requirements.txt
# Core RL and Environment
gymnasium==0.29.1
numpy>=1.26.0
pandas>=2.1.0

# Google Gemini API (FREE)
google-generativeai>=0.3.2

# Statistics and Analysis
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Jupyter support for Colab
ipykernel>=6.25.0
jupyter>=1.0.0